break Archive.py into itself and Find.py
